homework for the lesson 1, Level 2
