+++ src/level1/gui/Main.java
@@ -1,0 +1,12 @@
+package level1.gui;
+
+
+import javax.swing.*;
+
+public class Main {
+    public static void main(String[] args) {
+                MainFrame mainFrame = new MainFrame();
+                mainFrame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+                mainFrame.setVisible(true);
+    }
+}
