+++ src/level1/lessons/Lesson4.java
@@ -1,0 +1,126 @@
+package level1.lessons;
+
+import java.util.Random;
+import java.util.Scanner;
+
+public class Lesson4 {
+        public static int SIZE = 5;
+//    public static int SIZE = 3;
+    public static final char DOT_EMPTY = '•';
+    public static final char DOT_X = 'X';
+    public static final char DOT_O = 'O';
+    public static char[][] map;
+    public static Scanner sc = new Scanner(System.in);
+    public static Random rand = new Random();
+
+    public static void main(String[] args) {
+        initMap();
+        printMap();
+        while (true) {
+            humanTurn();
+            printMap();
+            if (checkWin(DOT_X)) {
+                System.out.println("Победил человек");
+                break;
+            }
+            if (isMapFull()) {
+                System.out.println("Ничья");
+                break;
+            }
+            aiTurn();
+            printMap();
+            if (checkWin(DOT_O)) {
+                System.out.println("Победил Искуственный Интеллект");
+                break;
+            }
+            if (isMapFull()) {
+                System.out.println("Ничья");
+                break;
+            }
+        }
+        System.out.println("Игра закончена");
+    }
+
+    public static boolean checkWin(char symb) {
+        int l = 0, n = 0;
+        for (int i = 0; i < SIZE; i++) {
+            int k = 0, f = 0;
+            for (int j = 0; j < SIZE; j++) {
+                if (map[i][j] == symb) k++;
+                if (map[j][i] == symb) f++;
+            }
+            if (map[i][i] == symb) l++;
+            if (map[SIZE - i - 1][i] == symb) n++;
+
+            if (k == SIZE || f == SIZE || l == SIZE || n == SIZE) return true;
+        }
+        return false;
+    }
+
+    public static boolean isMapFull() {
+        for (int i = 0; i < SIZE; i++) {
+            for (int j = 0; j < SIZE; j++) {
+                if (map[i][j] == DOT_EMPTY) return false;
+            }
+        }
+        return true;
+    }
+
+    public static void aiTurn() {
+        int x, y;
+        do {
+            x = getNextX(SIZE);
+            y = getNextY(SIZE);
+        } while (!isCellValid(x, y));
+        System.out.println("Компьютер походил в точку " + (x + 1) + " " + (y + 1));
+        map[y][x] = DOT_O;
+    }
+
+    static int getNextX(int size) {
+        return rand.nextInt(size);
+    }
+
+    static int getNextY(int size) {
+        return rand.nextInt(size);
+    }
+
+    public static void humanTurn() {
+        int x, y;
+        do {
+            System.out.println("Введите координаты в формате X Y");
+            x = sc.nextInt() - 1;
+            y = sc.nextInt() - 1;
+        } while (!isCellValid(x, y)); // while(isCellValid(x, y) == false)
+        map[y][x] = DOT_X;
+    }
+
+    public static boolean isCellValid(int x, int y) {
+        if (x < 0 || x >= SIZE || y < 0 || y >= SIZE) return false;
+        if (map[y][x] == DOT_EMPTY) return true;
+        return false;
+    }
+
+    public static void initMap() {
+        map = new char[SIZE][SIZE];
+        for (int i = 0; i < SIZE; i++) {
+            for (int j = 0; j < SIZE; j++) {
+                map[i][j] = DOT_EMPTY;
+            }
+        }
+    }
+
+    public static void printMap() {
+        for (int i = 0; i <= SIZE; i++) {
+            System.out.print(i + " ");
+        }
+        System.out.println();
+        for (int i = 0; i < SIZE; i++) {
+            System.out.print((i + 1) + " ");
+            for (int j = 0; j < SIZE; j++) {
+                System.out.print(map[i][j] + " ");
+            }
+            System.out.println();
+        }
+        System.out.println();
+    }
+}
