+++ src/level1/oop/advanced/behavior/JumpBehaviour.java
@@ -1,0 +1,12 @@
+package level1.oop.advanced.behavior;
+
+public class JumpBehaviour extends Behaviour {
+    public JumpBehaviour(int threshold) {
+        super(threshold);
+    }
+
+    @Override
+    protected boolean perform(int val) {
+        return val <= threshold;
+    }
+}
