+++ src/level2/lesson1/behavior/BarrierBehavior.java
@@ -1,0 +1,5 @@
+package level2.lesson1.behavior;
+
+public interface BarrierBehavior {
+    boolean passedBarrier(SportBehavior sportBehavior);
+}
