+++ src/level2/lesson1/behavior/SportBehavior.java
@@ -1,0 +1,9 @@
+package level2.lesson1.behavior;
+
+public interface SportBehavior {
+    boolean run(double value);
+
+    boolean jump(double val);
+
+    boolean getWinFlag();
+}
