--- src/oop/advanced/behavior/Behaviour.java
@@ -1,15 +1,0 @@
-package oop.advanced.behavior;
-
-public abstract class Behaviour {
-    protected int threshold;
-
-    public Behaviour(int threshold) {
-        this.threshold = threshold;
-    }
-
-    protected abstract boolean perform(int val);
-
-    public boolean behave(int val) {
-        return perform(val);
-    }
-}
